# unified-backend/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: WMS Backend Data Layer API
  description: |
    ### Member 3 — Data Layer API Service
    Provides high-performance CRUD endpoints, inventory queries, stock summaries, dashboard analytics, user management, and reference data for the Warehouse Management System.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local / EC2 Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/products:
    get:
      summary: List products with dynamic filters, pagination, and sorting
      tags: [Products]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: search
          schema: { type: string }
          description: Partial search on name or SKU
        - in: query
          name: category_id
          schema: { type: string, format: uuid }
        - in: query
          name: stock_status
          schema: { type: string, enum: [in_stock, low, out] }
        - in: query
          name: sort_by
          schema: { type: string, enum: [name, sku, stock_qty, created_at], default: created_at }
        - in: query
          name: sort_order
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        200:
          description: Paginated product list
    post:
      summary: Create a new product (INVENTORY_MANAGER)
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, sku]
              properties:
                name: { type: string, example: "Industrial Bolt M8" }
                sku: { type: string, example: "BLT-M8-001" }
                barcode: { type: string, example: "8909876543210" }
                category_id: { type: string, format: uuid }
                unit_of_measure: { type: string, default: "pcs" }
                tracking_type: { type: string, enum: [none, lot, serial], default: "none" }
                reorder_min: { type: integer, default: 0 }
                reorder_max: { type: integer, default: 0 }
                reorder_qty: { type: integer, default: 0 }
                image_url: { type: string, format: uri }
      responses:
        201:
          description: Product created successfully
  /api/v1/products/{id}:
    get:
      summary: Get single product detail + stock per location
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        200:
          description: Product detail with location stock breakdown
        404:
          description: Product not found
    patch:
      summary: Update product fields (INVENTORY_MANAGER)
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                reorder_min: { type: integer }
                reorder_max: { type: integer }
      responses:
        200:
          description: Product updated
    delete:
      summary: Soft delete product (ADMIN) - checks active stock
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        200:
          description: Product deactivated
        400:
          description: Cannot delete product with active stock
...